--- .changeset/lovely-planets-sort.md
@@ -1,5 +1,0 @@
----
-'@leafygreen-ui/typography': minor
----
-
-Adds OrderedList and ListItem components
