+++ packages/ordered-list/README.md
@@ -1,0 +1,32 @@
+# Ordered List
+
+![npm (scoped)](https://img.shields.io/npm/v/@leafygreen-ui/ordered-list.svg)
+
+#### [View on MongoDB.design](https://www.mongodb.design/component/ordered-list/example/)
+
+## Installation
+
+### Yarn
+
+```shell
+yarn add @leafygreen-ui/ordered-list
+```
+
+### NPM
+
+```shell
+npm install @leafygreen-ui/ordered-list
+```
+
+# OrderedList
+
+| Prop       | Type      | Description                                      | Default |
+| ---------- | --------- | ------------------------------------------------ | ------- |
+| `darkMode` | `boolean` | Determines if the component renders in dark mode | `false` |
+
+# ListItem
+
+| Prop        | Type              | Description                                                    | Default |
+| ----------- | ----------------- | -------------------------------------------------------------- | ------- |
+| title       | `string`          | The title of the step.                                         |         |
+| description | `React.ReactNode` | The description of the step. This will render below the title. |         |
