+++ packages/ordered-list/src/OrderedList/OrderedList.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import { ListItem } from '../ListItem';
+import { ListItem } from '../../ordered-list/src/OrderedList/ListItem';
 
 import { OrderedList } from './OrderedList';
