+++ packages/ordered-list/src/index.ts
@@ -1,5 +1,5 @@
-export { ListItem } from './ListItem';
 export { OrderedList, OrderedListProps } from './OrderedList';
+export { ListItem } from './ListItem';
 export {
   OrderedListContext,
   useOrderedListContext,
