--- packages/typography/src/OrderedList/ListItem/ListItem.types.ts
@@ -1,9 +1,0 @@
-import { HTMLElementProps } from '@leafygreen-ui/lib';
-
-import { CommonTypographyProps } from '../../types';
-
-export type ListItemProps = HTMLElementProps<'li'> &
-  CommonTypographyProps & {
-    title?: React.ReactNode;
-    description?: React.ReactNode;
-  };
